% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{UTF-8}
\name{dictionary_eng}
\alias{dictionary_eng}
\title{Dictionary for PNAD-Contínua microdata - English}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 947 rows and 5 columns.
}
\usage{
dictionary_eng
}
\description{
This dataframe contains information on PNAD-Contínua data. It describes,
in Portuguese, each variable, the possible values assumed by them as well
as its class (factor or double). All information is in English. Check \code{dictionary_pt.br}
for the version in Portuguese
}
\seealso{
\code{\link{dictionary_pt.br}}
}
\keyword{datasets}
