% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\encoding{UTF-8}
\name{dictionary_pt.br}
\alias{dictionary_pt.br}
\title{Dictionary for PNAD-Contínua microdata - Portuguese}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 947 rows and 5 columns.
}
\usage{
dictionary_pt.br
}
\description{
This dataframe contains information on PNAD-Contínua data. It describes,
in portuguese, each variable, the possible values assumed by them as well
as its class (factor or double). All information is in Portuguese. Check \code{dictionary_eng}
Prices of 50,000 round cut diamonds.
}
\seealso{
\code{\link{dictionary_eng}}
}
\keyword{datasets}
