% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrating_functions.R
\encoding{UTF-8}
\name{pnadc_panel}
\alias{pnadc_panel}
\title{Builds PNAD Contínua panel data from a loaded object}
\usage{
pnadc_panel(..., basic = TRUE, lang = "english")
}
\arguments{
\item{...}{PNAD Contínua dataframes to be considered in panel}

\item{basic}{Basic or advanced identification. Defaults to \code{TRUE}}

\item{lang}{Idiom used on the original data. Defaults to \code{english}. If in Portuguese,
set \code{land = 'portuguese'}.}
}
\value{
List of dataframes. Each element corresponds to a different
panel present in the original dataset.
}
\description{
Builds PNAD Contínua panel data from a loaded object
}
\examples{

\dontrun{
Let \code{data1} and \code{data2} be two dataframes with deidentified
PNAD-C data:

panel <- pnadc_panel(data1, data2, basic = TRUE, lang = 'english')

The user could also use a list of dataframes  \code{list_df}:

panel <- pnadc_panel(list_df, basic = TRUE)
}
}
